Added call to cluster finding program
Fixed a problem with the filename  having junk at the end and added in the calls to useful_fits and phase_fits_bank. When reference or measurement files are loaded, they are now run through the cluster fitting program and the combobox under the correct frame is updated with the cluster choices from each file. Next step will probably be to put the plots on the main window, then to add in the remaining  calculation steps and output the information to the GUI. Finally, adding in the ability to select different clusters of interest with the combobox.