Implemented plotting in GUI
Added the function of plotting the reference scatter plot directly into the GUI, including passing in the information about the clusters to the main program. Next step is to implement the same thing for the measurement, turn off the plotting in the called file, and then start actually analyzing. Oh, and I resized everything.