Calculations, part one
Added in a new file to call that takes the data from a loaded file and uses it to calculate the angle and radius of the spot of interest. Main file outputs that information to the boxes under the graphs. Also added in measurement file plotting and turned off the outside program's plotting. And added some comments.