Global variables added
The information from the reference and measurement plots needed to calculate angles and frequencies are now stored in global variables. When a different clusters is selected, these global variables are used to recalculate the angle and radius. Next step will be to use the global variables in the frequency calculations, rather than scraping the information from the GUI.