Global variables part 2
The program now uses the global variables when calculating the frequency. It also checks the selected cluster, so those can be changed before pressing enter. Next step will be making a widget to enter in calibration data and the species of interest and pulling in the AME text to calculate the estimated frequencies.  I also ned to figure out how to enter the accumulation time.