Widget window!
Added a file with code tocreate a new widget window to enter in calibration data. Oof. Next step will be to actually add functionality to the calibration window. Potentially also moving the GUI code for the main window to a new file to make the function code easier to look at.