Calibration Widget
Calibration Window fully functional. The cancel button closes the window. The Accept button rewrites the calibration file and the global calibration data variables, then closes the window.  Next step is to pull in the AME data, then figure out how to use that and the calibration data to find the expected cyclotron frequncy of the isotope of interest.