AME Data!
The GUI now uses the AME data file to find the mass of the calibration ion. It also recalculates the mass when the "accept" button is clicked in the calibration window. Next step is to use the calibration mass and frequency to find the expected frequency of the isotope of interest.