Expected Frequency Calculator!
Added a widget to input the nuclide of interest and the accumulation time. Both are now used to calculate the expected frequency of the isotope of interest to find the actual frequency. Also rewrites the "mass" file when the nuclide of interest is changed, pops up a window if the accumulation time isn't set, and prints the species of the isotope of interest on the main window.