Catching Errors
Added error catching for the most likely errors I could think of: missing files for calibration, mass, and AME data and forgetting to put in an accumulation time.  Maybe now will put in a "select new file" option for the calibration and mass files, but I think this is version 1.0.