Change N and detector center
Added buttons to the main UI to allow for changing N by one. Frequency recalculates when buttons are pressed. Also added a section to the calibration widget to allow for inputting the detector center. Reference and measurement information recalculates when the calibration information is changed. Also, it is now possible to hit the enter button when inputting information into the measurement or calibration widgets, rather than clicking a button. Created a new .exe file.